requirements/requirements.txt
numpy~=1.22.2
tqdm~=4.63.0
jax~=0.3.17
jaxlib~=0.3.14
setuptools~=58.0.4
neural-tangents~=0.6.0
torch~=1.13.0
scikit-learn~=1.0.2
matplotlib~=3.5.1
tensorflow~=2.9.2
torchvision~=0.14.0
scipy~=1.8.0
dataclasses~=0.6
cvxpy~=1.2.0
toma~=1.1.0
baal~=1.6.0
hydra-core~=1.2
omegaconf==2.2.3
pandas==1.5.3
dppy==0.3.2
colorlog
pynvml==11.4.1